Allow removal of malicious whitelist members
- Add a function that will allow the guardian (soon to be renamed to
  curator) to remove addresses from the whitelist.

- Assert that the recipient of a proposal is in the whitelist at
  `executeProposal()`

The thinking behind this patch is that a party of malicious token
holders can cooperate with a whitelist member and rob the DAO of its
money.

The token holders can still split away, but those who don't manage to
split in time will stay with a powerless guardian (curator) who will be
unable to stop the proposal from executing. So in this case the lazy
voters would lose all of their money.

By allowing the removal of malicious actors from the whitelist and by
double-checking that someone is still on the whitelist at the time of
`executeProposal()` the guardian (curator) can still save the lazy voters.