Force contractor to promise open source
An attempt to make explicit two things: (1) a promise from the contractor that the project developed under this proposal will be open source, and (2) an expectation (or a lack of an expectation) from the DAO that the materials will be open source.

Note that if the DAO sends 'false' to the sign function, they are saying to do not expect open source (but will accept it if offered). If the DAO sends 'true,' then the contract sill be signed only if the contractor has already promised to develop open source.

Forcing the contractor to develop open source (at the DAO's discretion) is important to insure that the product that is ultimately developed is exactly what the DAO is paying for. Trust but verify.